messaging_app/requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
Faker==37.4.0
inflection==0.5.1
packaging==25.0
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
mysqlclient
mysql-connector-python==8.2.0

